Restrict unauthenticated user

+++ base/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
@@ -10,6 +11,10 @@ from .forms import RoomForm
 
 
 def login_page(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -56,6 +61,8 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+# use this above the restricted view method, pass in the url to redirect the user
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -69,9 +76,13 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)  # find room
     form = RoomForm(instance=room)  # form will be prefilled with room values
+
+    if request.user != room.host:
+        return HttpResponse("You can't update the room")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -84,8 +95,12 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You can't delete the room")
 
     if request.method == 'POST':
         room.delete()
